Extract _version.py from _legacy.py
