+++ modules/helpers/__init__.py
@@ -43,4 +43,5 @@ from ._git import *  # noqa: F403
 from ._forms import *  # noqa: F403
 from ._os import *  # noqa: F403
 from ._schema import *  # noqa: F403
+from ._version import *  # noqa: F403
 from ._paths import *  # noqa: F403
