+++ modules/helpers/_version.py
@@ -1,0 +1,47 @@
+"""Version/branch detection utilities extracted from _legacy.py."""
+
+import os
+
+from git import Repo
+
+
+def get_remote_version(branch):
+    """Fetch the latest VERSION file from the correct GitHub branch."""
+    import requests
+
+    try:
+        response = requests.get(
+            f"https://raw.githubusercontent.com/Kometa-Team/Quickstart/{branch}/VERSION",
+            timeout=5,
+        )
+        response.raise_for_status()
+        version = response.text.strip()
+    except requests.RequestException:
+        return None  # If request fails, return None
+    try:
+        response = requests.get(
+            f"https://raw.githubusercontent.com/Kometa-Team/Quickstart/{branch}/BUILDNUM",
+            timeout=5,
+        )
+        response.raise_for_status()
+        build_num = response.text.strip()
+    except requests.RequestException:
+        build_num = "0"
+    return version if branch == "master" else f"{version}-build{build_num}"
+
+
+def get_branch():
+    # First priority: environment variable (Docker and CI use this)
+    branch = os.getenv("BRANCH_NAME")
+    if branch:
+        return branch
+
+    # Otherwise, try GitPython (if available)
+    if Repo:
+        try:
+            return Repo(path=".").head.ref.name  # noqa
+        except Exception:  # noqa
+            pass  # Ignore errors if GitPython fails
+
+    # Fallback: Use BRANCH_NAME from the environment (for non-Docker cases)
+    return os.getenv("BRANCH_NAME", "master")
